+++ cmd/k8s-operator/depaware.txt
@@ -1129,7 +1129,6 @@ tailscale.com/cmd/k8s-operator dependencies: (generated by github.com/tailscale/
    W    internal/syscall/windows                                     from crypto/internal/sysrand+
    W    internal/syscall/windows/registry                            from mime+
    W    internal/syscall/windows/sysdll                              from internal/syscall/windows+
-        internal/sysinfo                                             from testing
         internal/testlog                                             from os
         internal/trace/tracev2                                       from runtime+
         internal/unsafeheader                                        from internal/reflectlite+
@@ -1175,7 +1174,7 @@ tailscale.com/cmd/k8s-operator dependencies: (generated by github.com/tailscale/
         runtime/debug                                                from github.com/coder/websocket/internal/xsync+
         runtime/metrics                                              from github.com/prometheus/client_golang/prometheus+
         runtime/pprof                                                from net/http/pprof+
-        runtime/trace                                                from net/http/pprof+
+        runtime/trace                                                from net/http/pprof
         slices                                                       from encoding/base32+
         sort                                                         from compress/flate+
         strconv                                                      from compress/flate+
@@ -1184,7 +1183,6 @@ tailscale.com/cmd/k8s-operator dependencies: (generated by github.com/tailscale/
         sync                                                         from compress/flate+
         sync/atomic                                                  from context+
         syscall                                                      from crypto/internal/sysrand+
-        testing                                                      from tailscale.com/tka
         text/tabwriter                                               from k8s.io/apimachinery/pkg/util/diff+
         text/template                                                from html/template+
         text/template/parse                                          from html/template+
