+++ cmd/tailscaled/depaware-min.txt
@@ -10,11 +10,6 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
         github.com/go-json-experiment/json/internal/jsonwire         from github.com/go-json-experiment/json+
         github.com/go-json-experiment/json/jsontext                  from github.com/go-json-experiment/json+
         github.com/golang/groupcache/lru                             from tailscale.com/net/dnscache
-     💣 github.com/google/go-cmp/cmp                                 from tailscale.com/tka
-        github.com/google/go-cmp/cmp/internal/diff                   from github.com/google/go-cmp/cmp
-        github.com/google/go-cmp/cmp/internal/flags                  from github.com/google/go-cmp/cmp+
-        github.com/google/go-cmp/cmp/internal/function               from github.com/google/go-cmp/cmp
-     💣 github.com/google/go-cmp/cmp/internal/value                  from github.com/google/go-cmp/cmp
      💣 github.com/jsimonetti/rtnetlink                              from tailscale.com/net/netmon
         github.com/jsimonetti/rtnetlink/internal/unix                from github.com/jsimonetti/rtnetlink
         github.com/klauspost/compress                                from github.com/klauspost/compress/zstd
@@ -367,9 +362,8 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
         internal/synctest                                            from sync
         internal/syscall/execenv                                     from os+
         internal/syscall/unix                                        from crypto/internal/sysrand+
-        internal/sysinfo                                             from testing
         internal/testlog                                             from os
-        internal/trace/tracev2                                       from runtime+
+        internal/trace/tracev2                                       from runtime
         internal/unsafeheader                                        from internal/reflectlite+
         io                                                           from bufio+
         io/fs                                                        from crypto/x509+
@@ -401,11 +395,8 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
         path                                                         from io/fs+
         path/filepath                                                from crypto/x509+
         reflect                                                      from crypto/x509+
-        regexp                                                       from github.com/google/go-cmp/cmp+
-        regexp/syntax                                                from regexp
         runtime                                                      from crypto/internal/fips140+
         runtime/debug                                                from github.com/klauspost/compress/zstd+
-        runtime/trace                                                from testing
         slices                                                       from crypto/tls+
         sort                                                         from compress/flate+
         strconv                                                      from compress/flate+
@@ -413,7 +404,6 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
         sync                                                         from compress/flate+
         sync/atomic                                                  from context+
         syscall                                                      from crypto/internal/sysrand+
-        testing                                                      from tailscale.com/tka
         time                                                         from compress/gzip+
         unicode                                                      from bytes+
         unicode/utf16                                                from crypto/x509+
