+++ tsnet/depaware.txt
@@ -27,11 +27,6 @@ tailscale.com/tsnet dependencies: (generated by github.com/tailscale/depaware)
    L 💣 github.com/godbus/dbus/v5                                    from tailscale.com/net/dns
         github.com/golang/groupcache/lru                             from tailscale.com/net/dnscache
         github.com/google/btree                                      from gvisor.dev/gvisor/pkg/tcpip/header+
-     💣 github.com/google/go-cmp/cmp                                 from tailscale.com/tka
-        github.com/google/go-cmp/cmp/internal/diff                   from github.com/google/go-cmp/cmp
-        github.com/google/go-cmp/cmp/internal/flags                  from github.com/google/go-cmp/cmp+
-        github.com/google/go-cmp/cmp/internal/function               from github.com/google/go-cmp/cmp
-     💣 github.com/google/go-cmp/cmp/internal/value                  from github.com/google/go-cmp/cmp
   DI    github.com/google/uuid                                       from github.com/prometheus-community/pro-bing
         github.com/hdevalence/ed25519consensus                       from tailscale.com/tka
    L 💣 github.com/jsimonetti/rtnetlink                              from tailscale.com/net/netmon
@@ -466,7 +461,7 @@ tailscale.com/tsnet dependencies: (generated by github.com/tailscale/depaware)
         encoding/xml                                                 from github.com/tailscale/goupnp+
         errors                                                       from bufio+
         expvar                                                       from tailscale.com/health+
-        flag                                                         from tailscale.com/util/testenv+
+        flag                                                         from tailscale.com/util/testenv
         fmt                                                          from compress/flate+
         hash                                                         from crypto+
    W    hash/adler32                                                 from compress/zlib
@@ -518,7 +513,6 @@ tailscale.com/tsnet dependencies: (generated by github.com/tailscale/depaware)
    W    internal/syscall/windows                                     from crypto/internal/sysrand+
    W    internal/syscall/windows/registry                            from mime+
    W    internal/syscall/windows/sysdll                              from internal/syscall/windows+
-        internal/sysinfo                                             from testing
         internal/testlog                                             from os
         internal/trace/tracev2                                       from runtime+
         internal/unsafeheader                                        from internal/reflectlite+
@@ -559,7 +553,7 @@ tailscale.com/tsnet dependencies: (generated by github.com/tailscale/depaware)
         runtime                                                      from crypto/internal/fips140+
         runtime/debug                                                from github.com/coder/websocket/internal/xsync+
         runtime/pprof                                                from net/http/pprof+
-        runtime/trace                                                from net/http/pprof+
+ LDW    runtime/trace                                                from net/http/pprof
         slices                                                       from crypto/tls+
         sort                                                         from compress/flate+
         strconv                                                      from compress/flate+
@@ -568,7 +562,6 @@ tailscale.com/tsnet dependencies: (generated by github.com/tailscale/depaware)
         sync                                                         from compress/flate+
         sync/atomic                                                  from context+
         syscall                                                      from crypto/internal/sysrand+
-        testing                                                      from tailscale.com/tka
         text/tabwriter                                               from runtime/pprof
  LDW    text/template                                                from html/template
  LDW    text/template/parse                                          from html/template+
